pyproject.toml
[project]
name = "Fastapi-AI-Production-Boilerplate"
version = "0.3.0"
description = "Simple starter repo for your Machine Learning/AI projects"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "duckduckgo-search>=8.0.4",
    "fastapi-mcp>=0.3.4",
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-groq>=0.3.2",
    "langchain-tavily>=0.2.10",
    "langgraph>=0.5.3",
    "loguru>=0.7.3",
    "lxml[html-clean]>=5.4.0",
    "newspaper3k>=0.2.8",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.0",
    "scalar-fastapi>=1.2.2",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.34.2",
    "xgboost>=3.0.2",
]

[tool.uv.workspace]
members = [
    "test",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.12.0",
]
